+++ src/applications/gi-sandbox/components/content/modals/YellowRibbonModalContent.jsx
